+++ python/glide-async/python/glide/async_commands/ft.py
@@ -128,7 +128,7 @@ async def search(
     Returns:
         FtSearchResponse: A two element array, where first element is count of documents in result set, and the second
             element, which has the format Mapping[TEncodable, Mapping[TEncodable, TEncodable]] is a mapping between document
-            names and map of their attributes.
+            names and map of their attributes. When ``nocontent`` is set, the attribute maps will be empty.
         If count(option in `FtSearchOptions`) is set to true or limit(option in `FtSearchOptions`) is set to
             FtSearchLimit(0, 0), the command returns array with only one element - the count of the documents.