+++ python/glide-shared/glide_shared/commands/server_modules/ft_options/ft_constants.py
@@ -55,6 +55,8 @@ class FtSearchKeywords:
     LIMIT = "LIMIT"
     COUNT = "COUNT"
     AS = "AS"
+    NOCONTENT = "NOCONTENT"
+    DIALECT = "DIALECT"
 
 
 class FtAggregateKeywords: